Move subtensor rewrites from basic opt to subtensor opt
